feat: Node.__hash__
I was trying to use a Node object as a component of a cache key in https://github.com/ethereum/eth-abi/

Based on the class docstring, and the docstring for __repr__, it seems that these instances are suitable for hashing and this change should not cause any issues. Hopefully you agree, because using the repr in my cache key instead is not the solution I was hoping for! 😅